Renamed Auth Files & Created Login and SignUp Services
